+++ spec/models/form_submission_spec.rb
@@ -10,20 +10,21 @@ describe FormSubmission do
     end
   end
 
-  describe 'data_as_hash' do
-    it 'default' do
-      expect(build(klass).data_as_hash).to eq({})
+  describe 'data_object' do
+    it 'no results' do
+      expect(build(klass).data_object.foo).to be nil
     end
 
     it 'results' do
       submission = create(klass)
       submission.data.create!(name: 'foo', value: 'bar')
       submission.data.create!(name: 'baz', value: 'bak')
-      expect(submission.data_as_hash).to eq({ foo: 'bar', baz: 'bak' })
+      expect(submission.data_object.foo).to eq 'bar'
+      expect(submission.data_object.baz).to eq 'bak'
     end
   end
 
   it '#respond_to?' do
-    expect(build(klass)).to respond_to(:form, :data, :data_as_hash)
+    expect(build(klass)).to respond_to(:form, :data, :data_object)
   end
 end